requirements.txt.txt
websockets>=12.0
numpy>=1.24
pyserial>=3.5
pyserial-asyncio>=0.6
